# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WIP — Project Management API",
    "version": "1.0.0",
    "description": "REST API for WIP project management tool. Built with Django REST Framework."
  },
  "paths": {
...
    "/api/v1/tickets/{id}/": {
      "get": {
        "operationId": "tickets_retrieve",
        "description": "チケット詳細（コメント含む）",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "description": "A unique integer value identifying this チケット.",
            "required": true
          }
        ],
        "tags": [
          "tickets"
        ],
        "security": [
          {
            "jwtAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetail"
                }
              }
            },
            "description": ""
          }
        }
      },
      "put": {
        "operationId": "tickets_update",
        "description": "チケット更新",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "description": "A unique integer value identifying this チケット.",
            "required": true
          }
        ],
        "tags": [
          "tickets"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreateRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreateRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreateRequest"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "jwtAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketCreate"
                }
              }
            },
            "description": ""
          }
        }
      },
      "patch": {
        "operationId": "tickets_partial_update",
        "description": "チケットCRUD + コメント — ModelViewSet",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "description": "A unique integer value identifying this チケット.",
            "required": true
          }
        ],
        "tags": [
          "tickets"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchedTicketCreateRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/PatchedTicketCreateRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PatchedTicketCreateRequest"
              }
            }
          }
        },
        "security": [
          {
            "jwtAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketCreate"
                }
              }
            },
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "tickets_destroy",
        "description": "チケット削除",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "description": "A unique integer value identifying this チケット.",
            "required": true
          }
        ],
        "tags": [
          "tickets"
        ],
        "security": [
          {
            "jwtAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "No response body"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Category": {
        "type": "object",
        "description": "カテゴリー（読み取り用）",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "title": "名前",
            "maxLength": 100
          },
          "slug": {
            "type": "string",
            "title": "スラッグ",
            "description": "URLやフィルタ用の英字識別子",
            "maxLength": 50,
            "pattern": "^[-a-zA-Z0-9_]+$"
          },
          "level": {
            "allOf": [
              {
                "$ref": "#/components/schemas/LevelEnum"
              }
            ],
            "title": "階層レベル",
            "minimum": 0,
            "maximum": 32767
          },
          "parent": {
            "type": "integer",
            "nullable": true,
            "title": "親カテゴリー",
            "description": "level=2の場合、所属するlevel=1のカテゴリー"
          },
          "sort_order": {
            "type": "integer",
            "maximum": 2147483647,
            "minimum": 0,
            "title": "並び順"
          },
          "color": {
            "type": "string",
            "title": "表示カラー",
            "description": "HEXカラーコード（例: #6366f1）",
            "maxLength": 7
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "Comment": {
        "type": "object",
        "description": "コメント（読み取り用）",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "body": {
            "type": "string",
            "title": "本文"
          },
          "author": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserSummary"
              }
            ],
            "readOnly": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": [
          "author",
          "body",
          "createdAt",
          "id"
        ]
      },
...
      "LevelEnum": {
        "enum": [
          1,
          2
        ],
        "type": "integer",
        "description": "* `1` - 種別/フェーズ\n* `2` - カテゴリー"
      },
...
      "Milestone": {
        "type": "object",
        "description": "マイルストーン（読み取り用）",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "title": "名前",
            "maxLength": 200
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "readOnly": true
          },
          "description": {
            "type": "string",
            "title": "説明"
          },
          "project": {
            "type": "integer",
            "nullable": true,
            "title": "プロジェクト"
          },
          "openTicketCount": {
            "type": "integer",
            "readOnly": true
          },
          "closedTicketCount": {
            "type": "integer",
            "readOnly": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": [
          "closedTicketCount",
          "createdAt",
          "dueDate",
          "id",
          "name",
          "openTicketCount"
        ]
      },
...
      "PatchedTicketCreateRequest": {
        "type": "object",
        "description": "チケット（作成・更新用）",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "タイトル",
            "maxLength": 500
          },
          "description": {
            "type": "string",
            "title": "説明"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/StatusEnum"
              }
            ],
            "title": "ステータス"
          },
          "priority": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PriorityEnum"
              }
            ],
            "title": "優先度"
          },
          "ticket_type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketTypeEnum"
              }
            ],
            "title": "種別"
          },
          "assignee": {
            "type": "integer",
            "nullable": true,
            "title": "担当者"
          },
          "category": {
            "type": "integer",
            "nullable": true,
            "title": "カテゴリー",
            "description": "第2層カテゴリー（機能/コンポーネント）"
          },
          "project": {
            "type": "integer",
            "nullable": true,
            "title": "プロジェクト"
          },
          "milestone": {
            "type": "integer",
            "nullable": true,
            "title": "マイルストーン"
          },
          "parent": {
            "type": "integer",
            "nullable": true,
            "title": "親課題",
            "description": "子課題の場合、所属する親課題"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "title": "開始日"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "title": "期限日"
          }
        }
      },
      "PriorityEnum": {
        "enum": [
          "高",
          "中",
          "低"
        ],
        "type": "string",
        "description": "* `高` - 高\n* `中` - 中\n* `低` - 低"
      },
...
      "StatusEnum": {
        "enum": [
          "未対応",
          "処理中",
          "処理済み",
          "完了"
        ],
        "type": "string",
        "description": "* `未対応` - 未対応\n* `処理中` - 処理中\n* `処理済み` - 処理済み\n* `完了` - 完了"
      },
      "TicketCreate": {
        "type": "object",
        "description": "チケット（作成・更新用）",
        "properties": {
          "title": {
            "type": "string",
            "title": "タイトル",
            "maxLength": 500
          },
          "description": {
            "type": "string",
            "title": "説明"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/StatusEnum"
              }
            ],
            "title": "ステータス"
          },
          "priority": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PriorityEnum"
              }
            ],
            "title": "優先度"
          },
          "ticket_type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketTypeEnum"
              }
            ],
            "title": "種別"
          },
          "assignee": {
            "type": "integer",
            "nullable": true,
            "title": "担当者"
          },
          "category": {
            "type": "integer",
            "nullable": true,
            "title": "カテゴリー",
            "description": "第2層カテゴリー（機能/コンポーネント）"
          },
          "project": {
            "type": "integer",
            "nullable": true,
            "title": "プロジェクト"
          },
          "milestone": {
            "type": "integer",
            "nullable": true,
            "title": "マイルストーン"
          },
          "parent": {
            "type": "integer",
            "nullable": true,
            "title": "親課題",
            "description": "子課題の場合、所属する親課題"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "title": "開始日"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "title": "期限日"
          }
        },
        "required": [
          "title"
        ]
      },
      "TicketCreateRequest": {
        "type": "object",
        "description": "チケット（作成・更新用）",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "タイトル",
            "maxLength": 500
          },
          "description": {
            "type": "string",
            "title": "説明"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/StatusEnum"
              }
            ],
            "title": "ステータス"
          },
          "priority": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PriorityEnum"
              }
            ],
            "title": "優先度"
          },
          "ticket_type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketTypeEnum"
              }
            ],
            "title": "種別"
          },
          "assignee": {
            "type": "integer",
            "nullable": true,
            "title": "担当者"
          },
          "category": {
            "type": "integer",
            "nullable": true,
            "title": "カテゴリー",
            "description": "第2層カテゴリー（機能/コンポーネント）"
          },
          "project": {
            "type": "integer",
            "nullable": true,
            "title": "プロジェクト"
          },
          "milestone": {
            "type": "integer",
            "nullable": true,
            "title": "マイルストーン"
          },
          "parent": {
            "type": "integer",
            "nullable": true,
            "title": "親課題",
            "description": "子課題の場合、所属する親課題"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "title": "開始日"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "title": "期限日"
          }
        },
        "required": [
          "title"
        ]
      },
      "TicketDetail": {
        "type": "object",
        "description": "チケット詳細（コメント・添付ファイル含む）",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "ticketKey": {
            "type": "string",
            "readOnly": true
          },
          "title": {
            "type": "string",
            "title": "タイトル",
            "maxLength": 500
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/StatusEnum"
              }
            ],
            "title": "ステータス"
          },
          "priority": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PriorityEnum"
              }
            ],
            "title": "優先度"
          },
          "ticketType": {
            "type": "string",
            "readOnly": true
          },
          "assignee": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserSummary"
              }
            ],
            "readOnly": true
          },
          "author": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserSummary"
              }
            ],
            "readOnly": true
          },
          "category": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Category"
              }
            ],
            "readOnly": true
          },
          "milestone": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Milestone"
              }
            ],
            "readOnly": true
          },
          "project": {
            "type": "integer",
            "nullable": true,
            "title": "プロジェクト"
          },
          "parent": {
            "type": "integer",
            "nullable": true,
            "title": "親課題",
            "description": "子課題の場合、所属する親課題"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "readOnly": true
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "readOnly": true
          },
          "commentCount": {
            "type": "integer",
            "readOnly": true
          },
          "childCount": {
            "type": "integer",
            "readOnly": true
          },
          "gantt_order": {
            "type": "integer",
            "maximum": 2147483647,
            "minimum": -2147483648,
            "title": "ガント表示順",
            "description": "ガントチャートでの手動並べ替え順序"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "description": {
            "type": "string"
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            },
            "readOnly": true
          },
          "closedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": [
          "assignee",
          "author",
          "category",
          "childCount",
          "closedAt",
          "commentCount",
          "comments",
          "createdAt",
          "description",
          "dueDate",
          "id",
          "milestone",
          "startDate",
          "ticketKey",
          "ticketType",
          "title",
          "updatedAt"
        ]
      },
...
      "TicketTypeEnum": {
        "enum": [
          "不具合",
          "課題",
          "作業依頼",
          "QA"
        ],
        "type": "string",
        "description": "* `不具合` - 不具合\n* `課題` - 課題\n* `作業依頼` - 作業依頼\n* `QA` - QA"
      },
...
      "UserSummary": {
        "type": "object",
        "description": "ユーザー（FK参照用の軽量版）",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "username": {
            "type": "string",
            "title": "ユーザー名",
            "description": "この項目は必須です。半角アルファベット、半角数字、@/./+/-/_ で150文字以下にしてください。",
            "pattern": "^[\\w.@+-]+$",
            "maxLength": 150
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "メールアドレス",
            "maxLength": 254
          },
          "displayName": {
            "type": "string",
            "readOnly": true
          }
        },
        "required": [
          "displayName",
          "id",
          "username"
        ]
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "wip_sessionid"
      },
      "jwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}